Complex all over, need traits to be implimented

+++ src/output.rs
@@ -1,5 +1,7 @@
 use chrono::Local;
 use csv;
+use num_complex::Complex64;
+use num_traits::Zero;
 use ndarray::ArrayView3;
 use ordinal::Ordinal;
 use rayon;
@@ -30,9 +32,9 @@ struct ObservablesOutput {
     /// Excited state number.
     state: u8,
     /// Total energy.
-    energy: f64,
+    energy: Complex64,
     /// Binding energy.
-    binding_energy: f64,
+    binding_energy: Complex64,
     /// Coefficient of determination
     r: f64,
     /// Grid size / Coefficient of determination
@@ -49,7 +51,7 @@ struct PlainRecord {
     /// Index in *z*
     k: usize,
     /// Data at this position
-    data: f64,
+    data: Complex64,
 }
 
 /// Simply prints the Wafer banner with current commit info and thread count.
@@ -77,7 +79,7 @@ pub fn print_banner(sha: &str) {
 /// *`v` - The potential to output.
 /// * `project` - The project name (for directory to save to).
 /// * `file_type` - What type of file format to use in the output.
-pub fn potential(v: &ArrayView3<f64>, project: &str, file_type: &FileType) -> Result<()> {
+pub fn potential(v: &ArrayView3<Complex64>, project: &str, file_type: &FileType) -> Result<()> {
     let filename = format!(
         "{}/potential{}",
         get_project_dir(project),
@@ -96,7 +98,7 @@ pub fn potential(v: &ArrayView3<f64>, project: &str, file_type: &FileType) -> Re
 /// # Arguments
 /// *`array` - The array to output
 /// * `filename` - file / directory to save to.
-fn write_csv(array: &ArrayView3<f64>, filename: &str) -> Result<()> {
+fn write_csv(array: &ArrayView3<Complex64>, filename: &str) -> Result<()> {
     let mut buffer = csv::WriterBuilder::new()
         .has_headers(false)
         .from_path(filename)
@@ -120,7 +122,7 @@ fn write_csv(array: &ArrayView3<f64>, filename: &str) -> Result<()> {
 /// # Arguments
 /// *`array` - The array to output
 /// * `filename` - file / directory to save to.
-fn write_mpk(array: &ArrayView3<f64>, filename: &str) -> Result<()> {
+fn write_mpk(array: &ArrayView3<Complex64>, filename: &str) -> Result<()> {
     let mut output = Vec::new();
     array
         .serialize(&mut rmps::Serializer::new(&mut output))
@@ -138,7 +140,7 @@ fn write_mpk(array: &ArrayView3<f64>, filename: &str) -> Result<()> {
 /// # Arguments
 /// *`array` - The array to output
 /// * `filename` - file / directory to save to.
-fn write_json(array: &ArrayView3<f64>, filename: &str) -> Result<()> {
+fn write_json(array: &ArrayView3<Complex64>, filename: &str) -> Result<()> {
     let buffer = File::create(&filename)
         .chain_err(|| ErrorKind::CreateFile(filename.to_string()))?;
     serde_json::to_writer_pretty(buffer, array)
@@ -151,7 +153,7 @@ fn write_json(array: &ArrayView3<f64>, filename: &str) -> Result<()> {
 /// # Arguments
 /// *`array` - The array to output
 /// * `filename` - file / directory to save to.
-fn write_yaml(array: &ArrayView3<f64>, filename: &str) -> Result<()> {
+fn write_yaml(array: &ArrayView3<Complex64>, filename: &str) -> Result<()> {
     let buffer = File::create(&filename)
         .chain_err(|| ErrorKind::CreateFile(filename.to_string()))?;
     serde_yaml::to_writer(buffer, array)
@@ -170,7 +172,7 @@ fn write_yaml(array: &ArrayView3<f64>, filename: &str) -> Result<()> {
 /// * `project` - The project name (for directory to save to).
 /// * `file_type` - What type of file format to use in the output.
 pub fn wavefunction(
-    phi: &ArrayView3<f64>,
+    phi: &ArrayView3<Complex64>,
     num: u8,
     converged: bool,
     project: &str,
